unix: fix integer overflow in uv_hrtime
Conversion to nanoseconds was overflowing with 32-bit builds.

+++ src/unix/linux.c
@@ -134,7 +134,7 @@ static char* basename_r(const char* path) {
 uint64_t uv_hrtime() {
   struct timespec ts;
   clock_gettime(CLOCK_MONOTONIC, &ts);
-  return (ts.tv_sec * NANOSEC + ts.tv_nsec);
+  return (((uint64_t) ts.tv_sec) * NANOSEC + ts.tv_nsec);
 }
 
 void uv_loadavg(double avg[3]) {
